+++ testdata/de/dm/intellij/liferay/language/freemarker/LiferayFtlVariableProviderTest/com/liferay/portal/template/freemarker/internal/LiferayObjectConstructor.java
@@ -1,0 +1,11 @@
+package com.liferay.portal.template.freemarker.internal;
+
+import java.util.List;
+
+public class LiferayObjectConstructor {
+
+    public Object exec(List arguments) {
+        return null;
+    }
+
+}