+++ testdata/de/dm/intellij/liferay/language/freemarker/LiferayFtlVariableProviderTest/de/dm/MyObject.java
@@ -1,0 +1,9 @@
+package de.dm;
+
+public class MyObject {
+
+    public String getMyValue() {
+        return null;
+    }
+
+}