+++ changelog.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
 - Feature: Code completion for action/render/resource commands in JSP taglibs
 - Feature: Support for multiple OSGi service definitions in a @Component annotation
 - Feature: Inspections for service.xml
+- Feature: Class Name completion for objectUtil in Freemarker templates
 
 ## 0.0.3
 - Feature: Javascript injection for aui:validator and many onXXX attributes